+++ src/spicelib/analysis/dctran.c
@@ -289,7 +289,7 @@ DCtran(CKTcircuit *ckt,
      /* modeinittran set here */
         ckt->CKTag[0]=ckt->CKTag[1]=0;
         bcopy(ckt->CKTstate0, ckt->CKTstate1,
-          ckt->CKTnumStates*sizeof(double));
+              (size_t) ckt->CKTnumStates * sizeof(double));
 
 #ifdef WANT_SENSE2
         if(ckt->CKTsenInfo && (ckt->CKTsenInfo->SENmode & TRANSEN)){
