+++ src/xspice/cmpp/pp_lst.c
@@ -296,7 +296,7 @@ static Status_t read_modpath(
         if(n == 0)
             model = (Model_Info_t *) malloc(sizeof(Model_Info_t));
         else
-            model = (Model_Info_t *) realloc(model, (n + 1) * sizeof(Model_Info_t));
+            model = (Model_Info_t *) realloc(model, (size_t) (n + 1) * sizeof(Model_Info_t));
         model[n].path_name = NULL;
         model[n].spice_name = NULL;
         model[n].cfunc_name = NULL;
@@ -304,7 +304,7 @@ static Status_t read_modpath(
         model[n].cfunc_unique = TRUE;
 
         /* Put pathname into info structure */
-        model[n].path_name = (char *) malloc(len+1);
+        model[n].path_name = (char *) malloc((size_t) (len+1));
         strcpy(model[n].path_name, path);
 
         /* Increment count of paths read */
@@ -417,13 +417,13 @@ static Status_t read_udnpath(
         if(n == 0)
             node = (Node_Info_t *) malloc(sizeof(Node_Info_t));
         else
-            node = (Node_Info_t *) realloc(node, (n + 1) * sizeof(Node_Info_t));
+            node = (Node_Info_t *) realloc(node, (size_t) (n + 1) * sizeof(Node_Info_t));
         node[n].path_name = NULL;
         node[n].node_name = NULL;
         node[n].unique = TRUE;
 
         /* Put pathname into info structure */
-        node[n].path_name = (char *) malloc(len+1);
+        node[n].path_name = (char *) malloc((size_t) (len+1));
         strcpy(node[n].path_name, path);
 
         /* Increment count of paths read */
